Rename `native_sdk` to `oak_functions_native_sdk`
