+++ cc/oak_functions/native_sdk/symbols.map.ld
